refactor(cvtservice): extract fixture_generator.go from god-file
Extract GenerateFixture + convertResponse into a dedicated file. Pure
file movement, no behavior change.

Part 4 of 4 in the validator_service.go split (issue #107).
validator_service.go now 1016 LOC (was 1815).
